画像アップロード処理のリファクタリング
